+++ lib/src/reflection_factory_base.dart
@@ -21,7 +21,7 @@ import 'reflection_factory_utils.dart';
 /// Class with all registered reflections ([ClassReflection]).
 class ReflectionFactory {
   // ignore: constant_identifier_names
-  static const String VERSION = '2.2.3';
+  static const String VERSION = '2.2.4';
 
   static final ReflectionFactory _instance = ReflectionFactory._();
 
@@ -2506,8 +2506,8 @@ class TypeReflection<T> {
     }
 
     var typeStr = type.toString();
-    var idx = typeStr.indexOf('<');
-    if (idx > 0) typeStr = typeStr.substring(0, idx);
+    typeStr = TypeInfo.removeTypeGenerics(typeStr);
+
     return typeStr;
   }